+++ lib/rubyzen/declarations/block_declaration.rb
@@ -17,6 +17,8 @@ module Rubyzen
       include Rubyzen::Providers::RaisesProvider
       include Rubyzen::Providers::SourceCodeProvider
       include Rubyzen::Providers::CallSiteProvider
+      include Rubyzen::Providers::ReturnExpressionsProvider
+      include Rubyzen::Providers::AssignmentsProvider
 
       # @return [RuboCop::AST::Node]
       attr_reader :node